# server/openapi.yaml
openapi: 3.1.0
info:
  title: Climbing for Change API
  version: 1.0.0
  description: API for managing a charity climbing event, tracking teams, participants, and laps.
servers:
  - url: /api # Base path for all API endpoints
paths:
...
  /hills/{id}:
    get:
      tags:
        - Hills
      summary: Get a hill by its ID
      operationId: getHillById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the hill to retrieve.
      responses:
        '200':
          description: The requested hill.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HillPopulated'
        '404':
          description: Hill not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Hills
      summary: Update a hill by its ID
      operationId: updateHill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the hill to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HillRequest'
      responses:
        '200':
          description: Hill updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hill'
        '404':
          description: Hill not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Hills
      summary: Delete a hill by its ID
      operationId: deleteHill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the hill to delete.
      responses:
        '204':
          description: Hill deleted successfully. No content returned.
        '404':
          description: Hill not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /participants/{id}:
    get:
      tags:
        - Participants
      summary: Get a participant by their ID
      operationId: getParticipantById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the participant to retrieve.
      responses:
        '200':
          description: The requested participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantPopulated'
        '404':
          description: Participant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Participants
      summary: Update a participant by their ID
      operationId: updateParticipant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the participant to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantRequest'
      responses:
        '200':
          description: Participant updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '404':
          description: Participant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Participants
      summary: Delete a participant by their ID
      operationId: deleteParticipant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: The ID of the participant to delete.
      responses:
        '204':
          description: Participant deleted successfully. No content returned.
        '404':
          description: Participant not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        id: { type: string, description: Unique identifier for the location. }
        name: { type: string }
        address: { type: string }
        city: { type: string }
        provState: { type: string }
        country: { type: string }
        active: { type: boolean, default: true }
      required: [name, address, city, provState, country]
    Hill:
      type: object
      properties:
        id: { type: string, description: Unique identifier for the hill. }
        name: { type: string }
        location: { type: string, description: ID of the location where the hill exists. }
        lapElevationGain: { type: number, format: float }
        lapDistance: { type: number, format: float }
        elevationUnit: { type: string, enum: ['FT', 'M'] }
        distanceUnit: { type: string, enum: ['KM', 'MI'] }
      required: [name, location, lapElevationGain, lapDistance]
...
    Team:
      type: object
      properties:
        id: { type: string, description: Unique identifier for the team. }
        name: { type: string }
        event: { type: string, description: ID of the event the team participates in. }
        hill: { type: string, description: ID of the hill the team is climbing. }
        mountain: { type: string, description: ID of the virtual mountain the team chose. }
        rfidTagId: { type: string, nullable: true, description: ID of the RFID tag assigned to this team. }
        isSoloTeam: { type: boolean }
        lapsRequired: { type: number }
        totalDistanceRequired: { type: number }
        startDateTime: { type: string, format: date-time }
      required: [name, event, hill, mountain, lapsRequired, totalDistanceRequired, startDateTime]
    Participant:
      type: object
      properties:
        id: { type: string, description: Unique identifier for the participant. }
        firstName: { type: string }
        lastName: { type: string }
        teamId: { type: string, nullable: true, description: ID of the team the participant belongs to (can be null for a free agent). }
        active: { type: boolean }
      required: [firstName, lastName]
...
    ParticipantPopulated:
      allOf:
        - $ref: '#/components/schemas/Participant'
        - type: object
          properties:
            teamId: { $ref: '#/components/schemas/Team' }
...
    HillPopulated:
      allOf:
        - $ref: '#/components/schemas/Hill'
        - type: object
          properties:
            location: { $ref: '#/components/schemas/Location' }
            
    # --- Request Body Schemas (for POST/PUT) ---
...
    HillRequest:
      type: object
      properties:
        name: { type: string }
        location: { type: string, description: MongoDB ObjectId of the location. }
        lapElevationGain: { type: number }
        lapDistance: { type: number }
        elevationUnit: { type: string, enum: ['FT', 'M'] }
        distanceUnit: { type: string, enum: ['KM', 'MI'] }
      required: [name, location, lapElevationGain, lapDistance]
...
    ParticipantRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        teamId: { type: string, nullable: true, description: MongoDB ObjectId of the team. }
      required: [firstName, lastName]
...
    ErrorResponse:
      type: object
      properties: { error: { type: string } }
      required: [error]
      example: { error: "Resource not found" }
...